feature/#145392065/crud-operations-on-documents: edit code based on bug
