+++ client/components/includes/Footer.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+const Footer = () => {
+  return (
+    <div>
+    <footer className="page-footer light-reddish darken-3">
+      <div className="footer-copyright">
+        <div className="container">
+          <span>Copyright © 2015
+          <a className="grey-text text-lighten-4"
+          href="http://themeforest.net/user/geekslabs/portfolio?ref=geekslabs"
+          target="_blank">DataHubs</a> All rights reserved.</span>
+          <span className="right"> Project by
+          <a className="grey-text text-lighten-4"
+          href="http://github.com/andela-ksolomon">Solomon Kingsley</a></span>
+          </div>
+      </div>
+    </footer>
+    </div>
+  );
+};
+export default Footer;
