+++ client/index.html
@@ -4,6 +4,11 @@
   <title>DataHub</title>
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
    <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css" rel="stylesheet" type="text/css" media="screen" />
+   <style>
+   .page-footer {
+     display: none;
+   }
+   </style>
 </head>
 <body>
  <!-- START MAIN -->
@@ -18,17 +23,22 @@
     <!-- END CONTENT -->
   </div>
   </div>
-  <br/>
-  <br/>
-  <footer class="page-footer light-reddish darken-3">
+  <br />
+  <br />
+  <footer id="foot" class="page-footer light-reddish darken-3">
     <div class="footer-copyright">
       <div class="container">
-        <span>Copyright © 2015 <a class="grey-text text-lighten-4" href="http://themeforest.net/user/geekslabs/portfolio?ref=geekslabs" target="_blank">GeeksLabs</a> All rights reserved.</span>
-        <span class="right"> Project by <a class="grey-text text-lighten-4" href="http://github.com/andela-ksolomon">Solomon Kingsley</a></span>
+        <span>Copyright © 2015
+        <a class="grey-text text-lighten-4"
+        href="http://github.com/andela-ksolomon"
+        target="_blank">DataHubs</a> All rights reserved.</span>
+        <span class="right"> Project by
+        <a class="grey-text text-lighten-4"
+        href="http://github.com/andela-ksolomon">Solomon Kingsley</a></span>
         </div>
     </div>
   </footer>
-  </body>
+</body>
   <!-- END MAIN -->
    <script src="//tinymce.cachefly.net/4.2/tinymce.min.js"></script>
     <script type="text/javascript" src="bundle.js"></script>
